+++ familyhelper-finance-impl/src/main/java/com/dwarfeng/familyhelper/finance/impl/util/FtpConstants.java
@@ -8,7 +8,7 @@ package com.dwarfeng.familyhelper.finance.impl.util;
  */
 public final class FtpConstants {
 
-    public static final String PATH_BILL_FILE = "familyhelper-finance/bill-file";
+    public static final String[] PATH_BILL_FILE = {"familyhelper-finance", "bill-file"};
 
     private FtpConstants() {
         throw new IllegalStateException("禁止实例化");